+++ startup/gui/editor.py
@@ -33,7 +33,6 @@
 #  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 ##########################################################################
-import weakref
 
 import GafferUI
 import GafferSceneUI
@@ -42,6 +41,7 @@ def __editorCreated( editor ) :
 
 	GafferUI.GraphBookmarksUI.connectToEditor( editor )
 	GafferSceneUI.SceneHistoryUI.connectToEditor( editor )
+	GafferSceneUI.EditScopeUI.connectToEditor( editor )
 
 GafferUI.Editor.instanceCreatedSignal().connect( __editorCreated )
 GafferUI.CompoundEditor.nodeSetMenuSignal().connect( GafferUI.GraphBookmarksUI.appendNodeSetMenuDefinitions )